requirements.txt
discord.py 
random 
rich 
